Speedup parsing of really-long files.
Especially node spanning up really large amount of lines are taking
forever in particular:

    - The mapping line -> Node takes forever to be constructed, so
      replace it by a pseudo-interval-tree like data structure, which
      stores only start/end of nodes and is lazy/caching.
    - Don't iterate on all the lines a node spans.

From local profiling with example present in
https://github.com/ipython/ipython/issues/13731,
this make the construction of self._nodes_by_line be negligible in the
`__init__` function, vs taking ~50+ percent of the time.

That reduce my test example from 2.35s to 1.43s, with a interpreter
startup of ~0.25s.